requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-Bcrypt>=1.0.0
mysql-connector-python>=8.0.0
reportlab>=4.0.0
gunicorn==22.0.0
Werkzeug>=3.0.0
